fix(build): never guess the target database in the refresh tool
`refresh-email-template.ts` fiel ohne gesetzte Umgebungsvariablen auf
einen fest verdrahteten Verbindungsstring zurück und hätte damit still
in die falsche Datenbank geschrieben. In einem Git-Worktree fehlt die
`.env` regelmäßig — genau dort greift dieser Fall.

`cleanup-orphaned-uploads.ts` vermeidet das bereits bewusst und hatte
dafür ein getestetes `resolveConnectionString()`. Die Funktion liegt
jetzt samt `maskConnection()` in `dbConnection.ts` und wird von beiden
schreibenden Werkzeugen benutzt; `cleanup-orphaned-uploads.ts`
re-exportiert sie weiter für seinen Test.

Die Verbindung entsteht außerdem erst in `main()` statt beim Import:
`refresh-email-template.test.ts` importiert `decideTemplateRefresh` aus
der Datei und legte dabei bisher einen Verbindungspool an und las die
`.env`.
